Generalized watch-syslog's throttling method
Now it's possible to invoke the COMMAND after observing the end of a
series of PATTERN-like lines appearing in syslog with respect to a time
window, as well as invoking it as soon as those lines begin to appear.

HelloDisplays relies on watch-syslog's -e MSEC option.  It's because
Retina MacBook Pro produces more complex series of log lines than the
ordinary ones (e.g., 'WindowServer[#]: Display (added|removed)') when
the Retina display is mapped to a different resolution with external
displays.  Triggering HelloDisplays after all display-related events
were logged made more sense than anchoring on a specific line.